refactor: commenting out code in currently_unused dir
