+++ pkg/currently_unused/registries/registries_module_fn.go
@@ -1,21 +1,14 @@
 package registries
 
-import (
-	"github.com/gin-gonic/gin"
-	"github.com/sirupsen/logrus"
-	"github.com/techdecaf/k2s/v2/pkg/config"
-	"github.com/techdecaf/k2s/v2/pkg/kube"
-)
+// // Module registries module
+// func Module(app *gin.Engine, config *util.Config, k8s *kube.Service, logger *logrus.Entry) {
+// 	log := logger.WithFields(logrus.Fields{"module": "registries"})
 
-// Module registries module
-func Module(app *gin.Engine, config *config.ConfigService, k8s *kube.Service, logger *logrus.Entry) {
-	log := logger.WithFields(logrus.Fields{"module": "registries"})
+// 	registryService := &RegistryService{config: config, k8s: k8s, log: log}
 
-	registryService := &RegistryService{config: config, k8s: k8s, log: log}
+// 	if err := registryService.OnModuleInit(); err != nil {
+// 		log.Fatal(err)
+// 	}
 
-	if err := registryService.OnModuleInit(); err != nil {
-		log.Fatal(err)
-	}
-
-	log.Info("registries module loaded")
-}
+// 	log.Info("registries module loaded")
+// }
