+++ pkg/currently_unused/registries/registries_service_struct.go
@@ -1,98 +1,85 @@
 package registries
 
-import (
-	"strings"
+// // type RegistryServiceType interface {
+// // 	GetContainerRegistry() (namespaces *coreV1.Secret, err error)
+// // }
 
-	"github.com/reactivex/rxgo/v2"
-	"github.com/sirupsen/logrus"
-	"github.com/techdecaf/k2s/v2/pkg/config"
-	"github.com/techdecaf/k2s/v2/pkg/kube"
-	coreV1 "k8s.io/api/core/v1"
-	metaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	"k8s.io/apimachinery/pkg/watch"
-)
-
-// type RegistryServiceType interface {
-// 	GetContainerRegistry() (namespaces *coreV1.Secret, err error)
+// // RegistryService struct
+// type RegistryService struct {
+// 	config           *config.ConfigService
+// 	k8s              *kube.Service
+// 	log              *logrus.Entry
+// 	onNamespaceEvent rxgo.Observable
+// 	onNamespaceAdded rxgo.Observable
 // }
 
-// RegistryService struct
-type RegistryService struct {
-	config           *config.ConfigService
-	k8s              *kube.Service
-	log              *logrus.Entry
-	onNamespaceEvent rxgo.Observable
-	onNamespaceAdded rxgo.Observable
-}
+// // PrivateRegistry struct
 
-// PrivateRegistry struct
+// // ListPrivateRegisties method
+// func (t *RegistryService) ListPrivateRegisties() (registries *[]kube.PrivateRegistryDTO, err error) {
+// 	// registries, err := t.k8s.ListSecrets(t.config.SERVICE_NAME)
+// 	// metadata := []PrivateRegistryDTO{}
 
-// ListPrivateRegisties method
-func (t *RegistryService) ListPrivateRegisties() (registries *[]kube.PrivateRegistryDTO, err error) {
-	// registries, err := t.k8s.ListSecrets(t.config.SERVICE_NAME)
-	// metadata := []PrivateRegistryDTO{}
+// 	// for _, r := range registries.Items {
+// 	// 	if r.ObjectMeta.Labels["managed-by"] == t.config.SERVICE_NAME {
+// 	// 		metadata = append(metadata, PrivateRegistryDTO{
+// 	// 			Name:      r.ObjectMeta.Name,
+// 	// 			Namespace: r.ObjectMeta.Namespace,
+// 	// 			Labels:    r.ObjectMeta.Labels,
+// 	// 		})
+// 	// 	}
+// 	// }
 
-	// for _, r := range registries.Items {
-	// 	if r.ObjectMeta.Labels["managed-by"] == t.config.SERVICE_NAME {
-	// 		metadata = append(metadata, PrivateRegistryDTO{
-	// 			Name:      r.ObjectMeta.Name,
-	// 			Namespace: r.ObjectMeta.Namespace,
-	// 			Labels:    r.ObjectMeta.Labels,
-	// 		})
-	// 	}
-	// }
+// 	return registries, err
+// }
 
-	return registries, err
-}
+// // CopyPrivateRegistry method
+// func (t *RegistryService) CopyPrivateRegistry(stream rxgo.Observable) {
+// 	stream.DoOnError(func(err error) {
+// 		t.log.Error(err)
+// 	})
 
-// CopyPrivateRegistry method
-func (t *RegistryService) CopyPrivateRegistry(stream rxgo.Observable) {
-	stream.DoOnError(func(err error) {
-		t.log.Error(err)
-	})
+// 	for item := range stream.Observe() {
+// 		namespace := item.V.(*coreV1.Namespace)
+// 		if t.config.PRIVATE_REGISTRY_ENABLED == "true" && !strings.HasPrefix(namespace.Name, "kube-") {
+// 			t.log.Info("cloning private-registry to: " + namespace.Name)
 
-	for item := range stream.Observe() {
-		namespace := item.V.(*coreV1.Namespace)
-		if t.config.PRIVATE_REGISTRY_ENABLED == "true" && !strings.HasPrefix(namespace.Name, "kube-") {
-			t.log.Info("cloning private-registry to: " + namespace.Name)
+// 			if _, err := t.k8s.CopyRegistry("private-registry", t.config.SERVICE_NAME, namespace.Name); err != nil {
+// 				t.log.Error(err)
+// 			}
+// 		}
+// 	}
+// }
 
-			if _, err := t.k8s.CopyRegistry("private-registry", t.config.SERVICE_NAME, namespace.Name); err != nil {
-				t.log.Error(err)
-			}
-		}
-	}
-}
+// // OnModuleInit function description
+// func (t *RegistryService) OnModuleInit() error {
+// 	rx := kube.Rx{}
 
-// OnModuleInit function description
-func (t *RegistryService) OnModuleInit() error {
-	rx := kube.Rx{}
+// 	if t.config.PRIVATE_REGISTRY_ENABLED == "false" {
+// 		return nil
+// 	}
 
-	if t.config.PRIVATE_REGISTRY_ENABLED == "false" {
-		return nil
-	}
+// 	_, err := t.k8s.CreateRegistrySecret(&kube.CreateRegistryDTO{
+// 		Name:      "private-registry",
+// 		Namespace: t.config.SERVICE_NAME,
+// 		Registry:  t.config.PRIVATE_REGISTRY_URL,
+// 		Username:  t.config.PRIVATE_REGISTRY_USER,
+// 		Password:  t.config.PRIVATE_REGISTRY_PASS,
+// 	})
 
-	_, err := t.k8s.CreateRegistrySecret(&kube.CreateRegistryDTO{
-		Name:      "private-registry",
-		Namespace: t.config.SERVICE_NAME,
-		Registry:  t.config.PRIVATE_REGISTRY_URL,
-		Username:  t.config.PRIVATE_REGISTRY_USER,
-		Password:  t.config.PRIVATE_REGISTRY_PASS,
-	})
+// 	// listen to namespace events
+// 	t.onNamespaceEvent = t.k8s.OnNamespaceEvent(metaV1.ListOptions{
+// 		TypeMeta:            metaV1.TypeMeta{Kind: "Namespace"},
+// 		Watch:               true,
+// 		AllowWatchBookmarks: false,
+// 	})
 
-	// listen to namespace events
-	t.onNamespaceEvent = t.k8s.OnNamespaceEvent(metaV1.ListOptions{
-		TypeMeta:            metaV1.TypeMeta{Kind: "Namespace"},
-		Watch:               true,
-		AllowWatchBookmarks: false,
-	})
+// 	// filter on namespace added events
+// 	t.onNamespaceAdded = t.onNamespaceEvent.
+// 		Filter(rx.EventTypeFilter(watch.Added)).
+// 		Map(rx.NamespaceMap())
 
-	// filter on namespace added events
-	t.onNamespaceAdded = t.onNamespaceEvent.
-		Filter(rx.EventTypeFilter(watch.Added)).
-		Map(rx.NamespaceMap())
+// 	go t.CopyPrivateRegistry(t.onNamespaceAdded)
 
-	go t.CopyPrivateRegistry(t.onNamespaceAdded)
-
-	return err
-}
+// 	return err
+// }
